+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 2.4.3
-* Fix sync axis by computing min and max category values
+### Visual changes
 * Move "sync axis" and "show only main axis" settings to "Axis" card
 * Add font setttings for measure units and axis
+* Add an ‘auto width’ toggle for category labels that, when enabled, adjusts the label width to display all categories without trimming
+
+### Code improvements
+* Fix sync axis by computing min and max category values
 * When showing main axis only, reduce space between bars to save more place
 
 ## 2.4.2